Port Tease AI Fury
Decompiled the 3-5-2020 version of Fury and copied the changes. Readme updated to document a lot of Fury's changes. This includes decompiling and implementing the custom metronome (metro.dll) into the project. Also includes some fixes to bugs found during the porting process, and finally a bump to .net 4.8 for fixed blog images. More to come.